옵션 개별 적용 구현 중
