+++ public/data/CartData.json
@@ -2,6 +2,7 @@
   {
     "id": 1,
     "productName": "리무브",
+    "default": 1,
     "option": [
       { "value": 1 },
       { "value": 2 },
@@ -15,6 +16,7 @@
   {
     "id": 2,
     "productName": "컨트롤",
+    "default": 1,
     "option": [
       { "value": 1 },
       { "value": 2 },
@@ -28,6 +30,7 @@
   {
     "id": 3,
     "productName": "마스크",
+    "default": 1,
     "option": [
       { "value": 1 },
       { "value": 2 },
