+++ src/components/Nav/Cart/Cart.js
@@ -2,14 +2,15 @@ import React, { useState, useEffect } from 'react';
 import CartModal from './CartModal';
 
 function Cart() {
-  const [isCartModal, setIsCartModal] = useState(false);
+  // const [isClick, setIsClick] = useState(false);
+  const [modal, setModal] = useState('hidden');
   const handleModal = () => {
-    isCartModal ? setIsCartModal('hidden') : setIsCartModal('cartModal');
+    setModal('cartModal');
   };
   return (
     <div className="cart">
       <button onClick={handleModal}>카트</button>
-      <CartModal cartModal={isCartModal} />
+      <CartModal cartModal={modal} />
     </div>
   );
 }
