+++ src/pages/Main/Main.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Cart from '../../components/Nav/Cart';
+import Cart from '../../components/Nav/Cart/Cart';
 const Main = () => {
   return <Cart />;
 };
